Make DBManager a protocol and create EKDBManager
