+++ CardBox/ViewModels/Networking/HostGameManager.swift
@@ -10,7 +10,7 @@ protocol HostGameManager {
     var players: [String] { get }
     var gameRoomID: String { get }
     // TODO: make gameRunner be other gameRunners as well
-    var gameRunner: ExplodingKittensGameRunner? { get }
+    var gameRunner: GameRunnerProtocol? { get }
 
     func createRoom(player: Player)
     func removeFromRoom(player: Player)
